compute error

+++ core/include/structures.h
@@ -55,9 +55,12 @@ struct QuaternionPose
 
 struct Point3D
 {
-	double x;
-	double y;
-	double z;
+	//double x;
+	//double y;
+	//double z;
+	float x;
+	float y;
+	float z;
 	int index_pose;
 };
